temp code

+++ app/src/main/java/com/ftech/dev/android_my_food/ui/cart/CartViewModel.kt
@@ -5,25 +5,22 @@ import androidx.lifecycle.ViewModel
 import com.ftech.dev.android_my_food.data.model.Food
 import com.ftech.dev.android_my_food.data.model.Voucher
 import com.ftech.dev.android_my_food.data.repository.ItemInCartRepository
+import com.ftech.dev.android_my_food.data.repository.OrderRepository
 import com.ftech.dev.android_my_food.data.source.local.ItemInCartEntity
+import com.ftech.dev.android_my_food.data.source.local.OrderEntity
 import com.ftech.dev.android_my_food.data.source.local.SearchEntity
 
 class CartViewModel : ViewModel() {
 
     private val itemInCartRepository = ItemInCartRepository()
+    private val orderRepository = OrderRepository()
 
     val liveItemInCart = MutableLiveData<ItemInCartEntity>()
+    val liveOrders = orderRepository.getLiveDataOrders()
     val listItemInCartLiveData = itemInCartRepository.getAllItemInCartLiveData()
 
     var amount = MutableLiveData<Int>(0)
-
-//    private fun temp(): Int{
-//        var temp = 0
-//        for (item in listItemInCartLiveData.value!!){
-//            temp+= item.amount*item.total
-//        }
-//        return temp
-//    }
+    var isOrdering = MutableLiveData<Boolean>(true)
 
     var total = MutableLiveData<Int>(0)
     var promo = MutableLiveData<Int>(0)
@@ -45,9 +42,19 @@ class CartViewModel : ViewModel() {
     }
 
     fun deleteAll(){
-        amount.value = 0
-        total.value = 0
-        itemInCartRepository.deleteAll()
+        if (amount.value!=0) {
+            val orderEntity = OrderEntity(
+                name = "Order", total = total.value!!, paymentMethod = "Default",
+                amount = amount.value!!, date = "01/01/2021", status = true
+            )
+
+            insertOrder(orderEntity)
+
+            amount.value = 0
+            total.value = 0
+            itemInCartRepository.deleteAll()
+            isOrdering.value = false
+        }
     }
 
     fun insert(itemInCartEntity: ItemInCartEntity) {
@@ -56,5 +63,9 @@ class CartViewModel : ViewModel() {
 
     fun delete(itemInCartEntity: ItemInCartEntity) {
         itemInCartRepository.delete(itemInCartEntity)
+    }
+
+    fun insertOrder(orderEntity: OrderEntity){
+        orderRepository.insert(orderEntity)
     }
 }
